requirement.txt
functorch==0.1.1
imageio==2.22.0
lpips==0.1.4
matplotlib==3.6.0
opencv-python==4.6.0.66
Pillow==9.2.0
pyhocon==0.3.59
scikit-image==0.19.3
scipy==1.9.1
trimesh==3.15.1
chumpy==0.70
wandb==0.12.9
protobuf==3.20.0
numpy==1.23.4
pandas==1.5.1
